WIP #3551 - Add arguments for HostCommon#individual_puppetclasses
With the in_environment set to false (Default: true) Puppetclasses
will be returned even if not in the Host(Group) environment.

This can be used to identify missing classes when applying a new
environment, like in #3551, when changing the environment of a
HostGroup.

The additional with_environments allows to have the environments
included with the query via Puppetclass.with_possible_environments